Extract remix logic out of RouterPRovider into HydratedRouter
