+++ .changeset/create-remix-router.md
@@ -1,0 +1,5 @@
+---
+"react-router-dom": major
+---
+
+Use `createRemixRouter`/`RouterProvider` in `entry.client` instead of `RemixBrowser`
