+++ packages/remix-dev/config/defaults/entry.client.tsx
@@ -1,12 +1,12 @@
-import { RouterProvider } from "react-router-dom";
 import { startTransition, StrictMode } from "react";
 import { hydrateRoot } from "react-dom/client";
+import { HydratedRouter } from "react-router-dom";
 
 startTransition(() => {
   hydrateRoot(
     document,
     <StrictMode>
-      <RouterProvider />
+      <HydratedRouter />
     </StrictMode>
   );
 });
